refactor: rename blog components to PascalCase
